Cattle-Monitoring/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
opencv-python==4.8.1.78
numpy==1.26.4
pytest==8.0.0
python-dotenv==1.0.1